move buttons to popup, simplify styles
