+++ styles/popup.module.css
@@ -3,7 +3,7 @@
     flex-direction: column;
     padding: 16px;
     gap: 8px;
-    width: 150px;
+    width: 250px;
 }
 
 .flexRow {
