+++ styles/shared.module.css
@@ -3,20 +3,13 @@
     flex-direction: row;
     align-items: center;
     gap: 8px;
+    width: 100%;
 }
 
-.container {
+.flexCol {
     display: flex;
     flex-direction: column;
     align-items: center;
-    justify-content: space-between;
-    padding: 16px;
-    gap: 24px;
-}
-
-.content {
-    display: flex;
-    flex-direction: column;
     gap: 8px;
-    width: 400px;
+    width: 100%;
 }
